+++ jupiter-common/src/main/java/org/jupiter/common/concurrent/AffinityNamedThreadFactory.java
@@ -1,0 +1,87 @@
+/*
+ * Copyright (c) 2015 The Jupiter Project
+ *
+ * Licensed under the Apache License, version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at:
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jupiter.common.concurrent;
+
+import net.openhft.affinity.AffinityLock;
+import net.openhft.affinity.AffinityStrategies;
+import net.openhft.affinity.AffinityStrategy;
+import org.jupiter.common.util.internal.InternalThread;
+import org.jupiter.common.util.internal.logging.InternalLogger;
+import org.jupiter.common.util.internal.logging.InternalLoggerFactory;
+
+import java.util.concurrent.ThreadFactory;
+
+/**
+ * This is a ThreadFactory which assigns threads based the strategies provided.
+ * <p>
+ * If no strategies are provided AffinityStrategies.ANY is used.
+ *
+ * Jupiter
+ * org.jupiter.common.concurrent
+ *
+ * @author jiachun.fjc
+ */
+public class AffinityNamedThreadFactory implements ThreadFactory {
+
+    private static final InternalLogger logger = InternalLoggerFactory.getInstance(AffinityNamedThreadFactory.class);
+
+    private final String name;
+    private final boolean daemon;
+    private final AffinityStrategy[] strategies;
+    private AffinityLock lastAffinityLock = null;
+    private int id = 1;
+
+    public AffinityNamedThreadFactory(String name, AffinityStrategy... strategies) {
+        this(name, true, strategies);
+    }
+
+    public AffinityNamedThreadFactory(String name, boolean daemon, AffinityStrategy... strategies) {
+        this.name = name;
+        this.daemon = daemon;
+        this.strategies = strategies.length == 0 ? new AffinityStrategy[] { AffinityStrategies.ANY } : strategies;
+    }
+
+    @Override
+    public synchronized Thread newThread(final Runnable r) {
+        final String name2 = "affinity." + name + '-' + id++;
+        final Runnable r2 = new Runnable() {
+
+            @Override
+            public void run() {
+                AffinityLock al = lastAffinityLock == null ? AffinityLock.acquireLock() : lastAffinityLock.acquireLock(strategies);
+                try {
+                    if (al.cpuId() >= 0) {
+                        lastAffinityLock = al;
+                    }
+                    r.run();
+                } finally {
+                    al.release();
+                }
+            }
+        };
+        Thread t = new InternalThread(r2, name2);
+        try {
+            if (t.isDaemon() != daemon) {
+                t.setDaemon(daemon);
+            }
+        } catch (Exception ignored) { /* doesn't matter even if failed to set. */ }
+
+        logger.debug("Creates new {}.", t);
+
+        return t;
+    }
+}