+++ jupiter-example/src/main/java/org/jupiter/benchmark/tcp/BenchmarkClient.java
@@ -77,7 +77,13 @@ public class BenchmarkClient {
         SystemPropertyUtil.setProperty("jupiter.tracing.needed", "false");
         SystemPropertyUtil.setProperty("jupiter.use.non_blocking_hash", "true");
 
-        JClient client = new DefaultClient().withConnector(new JNettyTcpConnector(processors + 1));
+        JClient client = new DefaultClient().withConnector(new JNettyTcpConnector(processors + 1) {
+
+//            @Override
+//            protected ThreadFactory workerThreadFactory(String name) {
+//                return new AffinityNettyThreadFactory(name, AffinityStrategies.DIFFERENT_CORE);
+//            }
+        });
         client.connector().config().setOption(JOption.WRITE_BUFFER_HIGH_WATER_MARK, 512 * 1024);
         client.connector().config().setOption(JOption.WRITE_BUFFER_LOW_WATER_MARK, 256 * 1024);