+++ release-notes.md
@@ -1,6 +1,10 @@
 Jupiter release notes
 ------------------------
 
+### 2017-05-19:
+
+- [Thread Affinity](https://github.com/OpenHFT/Java-Thread-Affinity)
+
 ### 2017-05-19: version 1.2.5
 
 - 存在addAndGet场景时, 不使用jupiter自带的Atomic*FieldUpdater, 使用jdk默认的实现, 1.8后续版本jdk使用了intrinsics后更快